+++ nucleus/payara-modules/hazelcast-bootstrap/src/main/java/fish/payara/nucleus/hazelcast/contextproxy/CompleteConfigurationProxy.java
@@ -37,8 +37,9 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package fish.payara.nucleus.hazelcast;
+package fish.payara.nucleus.hazelcast.contextproxy;
 
+import fish.payara.nucleus.hazelcast.JavaEEContextUtil;
 import fish.payara.nucleus.hazelcast.JavaEEContextUtil.Context;
 import fish.payara.nucleus.hazelcast.JavaEEContextUtil.RequestContext;
 import java.util.Collection;
